Move IPython.frontend.zmqterminal to IPython.frontend.terminal.console
