+++ IPython/frontend/terminal/console/completer.py
